+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+* [Fix] Fix leaks and memory bugs in `tls_serialize_cert_chain`, `tls_free_serialized_chain`. Add tests.
+
 ## 0.90.12
 
 * [Fix] Change signature of some exported functions to be more consistent with the rest and simpler for C# bindings.